fix tests after review
