+++ spring-homework-3/src/test/java/ru/otus/spring/service/CsvFileReaderTest.java
@@ -1,5 +1,6 @@
 package ru.otus.spring.service;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -11,12 +12,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static ru.otus.spring.base.DefaultConstant.DEFAULT_QUESTIONS_NUMBER;
 
+@DisplayName("Сервис CsvFileReader должен")
 @SpringBootTest
 class CsvFileReaderTest {
 
     @Autowired
     private CsvFileReader fileReader;
 
+    @DisplayName("считать вопросы из csv файла.")
     @Test
     void readFile() {
         List<Question> questions = fileReader.readFile();
