+++ spring-homework-3/src/test/java/ru/otus/spring/service/OutputFormatterTest.java
@@ -1,5 +1,6 @@
 package ru.otus.spring.service;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -15,15 +16,17 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static ru.otus.spring.base.DefaultConstant.DEFAULT_QUESTIONS_NUMBER;
 import static ru.otus.spring.base.DefaultConstant.DEFAULT_TEST_RESULT;
 
+@DisplayName("Сервис OutputFormatter должен")
 @SpringBootTest
-class OutputFormatterImplTest {
+class OutputFormatterTest {
     private static final String DEFAULT_ANSWER_VALUE_1 = "value1";
     private static final String DEFAULT_ANSWER_VALUE_2 = "value2";
     private static final Boolean DEFAULT_ANSWER_IS_RIGHT = true;
 
     @Autowired
-    private OutputFormatterImpl outputFormatter;
+    private OutputFormatter outputFormatter;
 
+    @DisplayName("форматировать вопрос.")
     @Test
     void formatQuestion() {
         List<Answer> answerList = new ArrayList<>();
@@ -33,6 +36,7 @@ class OutputFormatterImplTest {
         assertEquals("question\n1)value1  2)value2  ", outputFormatter.formatQuestion(question));
     }
 
+    @DisplayName("форматировать результат")
     @Test
     void formatResult() {
         TestResult testResult = TestResult.builder()
